pyproject.toml
[project]
name = "llamaindex-ambientmeta"
version = "0.1.0"
description = "LlamaIndex integration for AmbientMeta Privacy Gateway — protect PII in LLM calls"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "AmbientMeta", email = "support@ambientmeta.com"},
]
keywords = ["llamaindex", "llama-index", "privacy", "pii", "llm", "ambientmeta", "rag", "sanitize"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "ambientmeta>=0.1.0",
    "llama-index-core>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://ambientmeta.com"
Documentation = "https://docs.ambientmeta.com/guides/llamaindex"
Repository = "https://github.com/AmbientMeta/llamaindex-ambientmeta"
Changelog = "https://github.com/AmbientMeta/llamaindex-ambientmeta/blob/main/CHANGELOG.md"
Issues = "https://github.com/AmbientMeta/llamaindex-ambientmeta/issues"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
llamaindex_ambientmeta = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 100
